requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
idna==2.9
more-itertools==8.2.0
multidict==4.7.5
packaging==20.3
Pillow==7.1.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytest==5.4.1
six==1.14.0
urllib3==1.25.9
wcwidth==0.1.9
yarl==1.4.2
